UDOO-mmcblk
Branch for installing on the soldered-in 'mmcblk' memory on the UDOO x86_64 single board computer.
https://www.udoo.org/docs-x86II/Introduction/Introduction.html

I use this device with two disks. I want a backup OS on the 'mmcblk' memory.
However since I use disk and parition identification by labels,
there must never be two same partition labels in one system.

Let's append '-MMCBLK' to the parition labels for the installation on the 'mmcblk' memory.
Capital letters has been chosen to keep consistency, as the EFI partition label seems to be only allowed in upper-case.

I also added the appendix to the hostname, to be able to easily differentiate between the OS instllations on different disks.