[GRUB] Refactor the video payload section
NOTE:
I've found out, that the UEFI systems have great problems with the
graphical interpretation of GRUB. It have insuferable latency,
described e.g. here:
  https://askubuntu.com/questions/1227735/grub-is-extremely-slow-1-second-per-key-input

I've been experimenting around and I foudn out,
that I can utilise the fact that every BLS_CFG calls the 'load_video' function
in such a way, that when the GRUB is started in graphical mode, it could be
switched to console mode right before the system starts to boot.

For some reason it doesn't work the other way around though.
Which would be the desirable setup - having GRUB in console mode, with no latency,
but having the boot in graphical mode.